+++ doc/ChangeLog.md
@@ -4,6 +4,8 @@
 
 - Migrated from CentOS 7 to Ubuntu Server 26.04 LTS
 
+- Added GitHub repository commit crawler.
+
 - Used MOEX and Bitpay instead of RBC.ru for quotes.
 
 - Bot admins can solve CAPTCHA instead of users now.
@@ -32,6 +34,7 @@
     - Spring Boot 2.6.7 => 4.1.1
     - Spring Boot Dependency Management 1.0.11.RELEASE => 1.1.7
     - com.github.pengrad:java-telegram-bot-api 5.7.0 => 10.1.0
+    - org.kohsuke:github-api => 2.0-rc.7
     - org.yaml:snakeyaml 1.29 => 2.6
     - org.jsoup:jsoup 1.14.3 => 1.23.1
     - org.nibor.autolink:autolink 0.10.0 => 0.12.0